Pause Menu Update/Add Vol MainMenu
I updated the Prefab for the Pause Menu to have an invisible back plate, so players couldn't accidentally hit next while in the pause menu, and adjusted colors to make buttons more readable.

Also disabled pause menu on Main Menu and added Game Volume to the lower left-hand side, since the pause menu is redundant in that scene.